pyproject.toml
[tool.poetry]
name = "veanu"
version = "0.1.3"
description = "veanu enables developers to access tools applicable to life sciences and healthcare."
authors = [
    "Domingo Bertus <hello@veanu.ch>",
    "NicoBerto <hello@veanu.ch>",
    "Penguin Berthold <hello@veanu.ch>",
]
repository = "https://github.com/open-veanu/veanu"
license = "MIT"
keywords = ["healthcare", "government", "opensource", "data science"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gradio = "^5.11.0"
build = "^1.2.2.post1"
twine = "^5.1.1"
abydos = "^0.5.0"
cologne-phonetics = "^2.0.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
bandit = "^1.7.10"
dacite = "^1.8.1"
pytest = "^8.3.4"
aiohttp = "^3.11.10"
pymupdf = "^1.25.1"
beautifulsoup4 = "^4.12.3"
openai = "^1.59.5"
defusedxml = "^0.7.1"
selenium = "^4.27.1"
torch = "^2.5.1"
transformers = "^4.46.3"
plotly = "^5.24.1"
llvmlite = "^0.43.0"
numba = "^0.60.0"
shap = "^0.46.0"
pytest-cov = "^6.0.0"
matplotlib = "^3.9.3"
sqlitecloud = "^0.0.83"
dateparser = "^1.2.0"
pdfminer = "^20191125"
google-search-results = "^2.4.2"

[tool.poetry.scripts]
veanu_drugsafetycompare_app = "veanu.drugsafetycompare.launch_demo_app:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools]
exclude-patterns = [".env"]